First commit
Contract Structure: The contract Lottery is defined with a state variable manager to store the address of the contract deployer and an array players to store the addresses of participants.

Constructor: The constructor initializes the manager variable with the address of the contract deployer.

alreadyEntered Function: This internal function checks whether a player has already entered the lottery by iterating through the players array.

enter Function: Players can call this function to enter the lottery by sending ether (msg.value) to the contract. It checks conditions such as whether the manager is trying to participate, whether the player has already entered, and whether the minimum amount of ether is sent.

random Function: This internal function generates a pseudo-random number using the sha256 hash function based on the current block's difficulty, block number, and the addresses of participants in the players array.

pickWinner Function: Only the manager can call this function. It generates a random index within the range of the players array, transfers the contract's balance to the winner, and resets the players array to start a new lottery.

getPlayer Function: This function allows anyone to retrieve the list of players who have participated in the lottery.